moved kb to az_search
