+++ requirements.txt
@@ -25,10 +25,6 @@ certifi==2023.7.22
 cffi==1.15.1
 chardet==5.2.0
 charset-normalizer==2.1.1
-chroma-bullet==2.2.0
-chroma-hnswlib==0.7.3
-chroma-migrate==0.0.7
-chromadb==0.4.17
 click==8.1.6
 clickhouse-connect==0.6.6
 coloredlogs==15.0.1
@@ -85,16 +81,6 @@ jsonpointer==3.0.0
 jupyter_client==8.3.0
 jupyter_core==5.3.1
 kubernetes==28.1.0
-langchain==0.3.1
-langchain-community==0.3.1
-langchain-core==0.3.6
-langchain-text-splitters==0.3.0
-langsmith==0.1.129
-llama-index-core==0.11.14
-llama-index-embeddings-azure-openai==0.2.5
-llama-index-embeddings-openai==0.2.5
-llama-index-llms-azure-openai==0.2.1
-llama-index-llms-openai==0.2.9
 lxml==4.9.3
 lz4==4.3.2
 MarkupSafe==2.1.3
