+++ src/database/__init__.py
@@ -1,12 +1,6 @@
-import sys
-__import__("pysqlite3")
-sys.modules["sqlite3"] = sys.modules.pop("pysqlite3")
-sys.path.append("src")
-
 from .base import BaseDB
 from .user_db import UserDB
 from .user_conv_db import UserConvDB
 from .bot_conv_db import BotConvDB
 from .expert_conv_db import ExpertConvDB
 from .user_relation_db import UserRelationDB
-from .faq_db import FAQDB
