+++ src/hierarchical_rag.py
@@ -1,11 +1,8 @@
 from datetime import datetime
 import json
-from embeddings.chroma.llama_index_azure_openi import get_chroma_llama_index_azure_openai_embeddings_fn
 from embeddings.chroma.openai import get_chroma_openai_embedding_fn
 from utils import get_client_with_token_provider, get_client_with_key
 import os
-import chromadb
-from chromadb.config import Settings
 import yaml
 
 with open("config.yaml") as file:
@@ -20,41 +17,27 @@ llm_client = get_client_with_key()
 model = os.environ["OPENAI_API_MODEL"].strip()
 general = "Generic"
 
-def hierarchical_rag_retrieve(query, org_id, num_chunks=3):
-    chroma_client = chromadb.PersistentClient(
-        path=persist_directory, settings=Settings(anonymized_telemetry=False)
-    )
-    collection = chroma_client.get_collection(
-        name=config["PROJECT_NAME"], embedding_function=embedding_fn
-    )
-    collection_count = collection.count()
-    print('collection ids count: ', collection_count)
-    relevant_chunks = collection.query(
-        query_texts=[query],
-        n_results=num_chunks,
-        where={"org_id": {"$in": [org_id, general]}}
-    )
-    citations: str = "\n".join(
-        [metadata["org_id"] + '-' + metadata["source"] for metadata in relevant_chunks["metadatas"][0]]
-    )
-
+def hierarchical_rag_retrieve(kb_client, query, org_id, num_chunks=3):
+    retrieved_chunks = kb_client.hybrid_search(query, org_id, num_chunks)
     relevant_chunks_string = ""
     relevant_update_chunks_string = ""
     chunks = []
 
     chunk1 = 0
     chunk2 = 0
-    for chunk, chunk_text in enumerate(relevant_chunks["documents"][0]):
-        if "kb_update" in relevant_chunks["metadatas"][0][chunk]["source"].lower():
-            relevant_update_chunks_string += (
-                f"Chunk #{chunk2 + 1}\n{chunk_text}\n\n"
-            )
-            chunk2 += 1
-            chunks.append((chunk_text, relevant_chunks["metadatas"][0][chunk]["source"].strip(), relevant_chunks["metadatas"][0][chunk]["org_id"].strip()))
+    for i, chunk in enumerate(retrieved_chunks):
+        if "kb_update" in chunk['metadata']['source'].lower():
+            relevant_update_chunks_string += (f"Chunk #{chunk2 + 1}\n{chunk['data_chunk']}\n\n")
+            chunk2 += 1  
         else:
-            relevant_chunks_string += f"Chunk #{chunk1 + 1}\n{chunk_text}\n\n"
+            relevant_chunks_string += f"Chunk #{chunk1 + 1}\n{chunk['data_chunk']}\n\n"
             chunk1 += 1
-            chunks.append((chunk_text, relevant_chunks["metadatas"][0][chunk]["source"].strip(), relevant_chunks["metadatas"][0][chunk]["org_id"].strip()))
+        chunks.append((chunk["data_chunk"], chunk["metadata"]["source"].strip(), chunk["org_id"].strip()))
+    
+    citations: str = "\n".join(
+        [ chunk["org_id"] + '-' + chunk["metadata"]["source"] for chunk in retrieved_chunks ]
+    )
+    
     return relevant_chunks_string, relevant_update_chunks_string, citations, chunks
 
 def hierarchical_rag_augment(retrieved_chunks, system_prompt, query, query_type, user_context):
